# src/main/resources/static/swagger.yaml
openapi: 3.0.1
info:
  title: Mannatime API
  description: 언제만나(mannatime) API Docs
  contact:
    email: apiteam@swagger.io
  version: 1.0.0
...
servers:
  - url: https://dev.swm-bentley-server.link
...
paths:
  /room:
    post:
      tags:
        - Room
      summary: 새로운 방 생성
      operationId: createRoom
      requestBody:
        $ref: 'requestBodies/roomRequest.yaml'
      responses:
        400:
          $ref: 'errors/room400Error.yaml'
  /room/{roomUuid}:
    get:
      tags:
        - Room
      summary: 방 정보 불러오기
      operationId: getRoomInfo
      parameters:
        - $ref: 'parameters/roomUuid.yaml'
      responses:
        200:
          $ref: 'responses/roomResponse.yaml'
        404:
          $ref: 'errors/room404Error.yaml'
...